requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0

# Diffusion Models
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# LoRA Training
peft>=0.7.0
bitsandbytes>=0.41.0

# Gradio UI
gradio>=4.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
tqdm>=4.66.0
omegaconf>=2.3.0
huggingface-hub>=0.19.0

# Dataset handling
datasets>=2.15.0

# Optimization
xformers>=0.0.22
triton>=2.1.0

# Monitoring (optional)
tensorboard>=2.15.0
wandb>=0.16.0
